% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ratio.R
\name{lr_minus}
\alias{lr_minus}
\title{Calculate negative likelihood ratio}
\usage{
lr_minus(sensitivity, specificity)
}
\arguments{
\item{sensitivity}{A numeric value representing sensitivity}

\item{specificity}{A numeric value representing specificity}
}
\value{
An object of class "lr_minus" with the calculated negative likelihood ratio and additional attributes
}
\description{
This function calculates the negative likelihood ratio based on sensitivity and specificity.
}
