% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ratio.R
\name{lr_plus}
\alias{lr_plus}
\title{Calculate positive likelihood ratio}
\usage{
lr_plus(sensitivity, specificity)
}
\arguments{
\item{sensitivity}{A numeric value representing sensitivity}

\item{specificity}{A numeric value representing specificity}
}
\value{
An object of class "lr_plus" with the calculated positive likelihood ratio and additional attributes
}
\description{
This function calculates the positive likelihood ratio based on sensitivity and specificity.
}
