% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pie_chart.R
\name{pie_chart}
\alias{pie_chart}
\title{Generate a pie chart with values and labels}
\usage{
pie_chart(Values, labels, col = NULL)
}
\arguments{
\item{Values}{A numeric vector of values to be plotted}

\item{labels}{A character vector of labels for each value}

\item{col}{vector of colors for the pie chart (optional)}
}
\value{
A pie chart of the input values with corresponding labels
}
\description{
Generate a pie chart with values and labels
}
