% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precision.R
\name{precision}
\alias{precision}
\title{Calculate precision}
\usage{
precision(true_positives, false_positives)
}
\arguments{
\item{true_positives}{A numeric value representing the number of true positives}

\item{false_positives}{A numeric value representing the number of false positives}
}
\value{
An object of class "precision" with the calculated precision and additional attributes
}
\description{
This function calculates the precision based on the number of true positives and false positives.
}
