% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recall.R
\name{recall}
\alias{recall}
\title{Calculate recall}
\usage{
recall(true_positives, false_negatives)
}
\arguments{
\item{true_positives}{A numeric value representing the number of true positives}

\item{false_negatives}{A numeric value representing the number of false negatives}
}
\value{
An object of class "recall" with the calculated recall and additional attributes
}
\description{
This function calculates the recall based on the number of true positives and false negatives.
}
