% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ability.R
\name{Test_Ability}
\alias{Test_Ability}
\title{Function to calculate test ability}
\usage{
Test_Ability(TP, FP, TN, FN)
}
\arguments{
\item{TP}{numeric value representing the number of true positive test results}

\item{FP}{numeric value representing the number of false positive test results}

\item{TN}{numeric value representing the number of true negative test results}

\item{FN}{numeric value representing the number of false negative test results}
}
\value{
An object of class "test_results" containing the test ability results
}
\description{
This function takes in four numeric values representing the number of true positive (TP), false positive (FP), true negative (TN), and false negative (FN) test results, and returns an object of class "test_results" containing the test ability results.
}
