% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ability.R
\name{test_results}
\alias{test_results}
\title{Function to create test results object}
\usage{
test_results(Sensitivity, Specificity, PPV, NPV)
}
\arguments{
\item{Sensitivity}{numeric value representing the sensitivity of the test}

\item{Specificity}{numeric value representing the specificity of the test}

\item{PPV}{numeric value representing the positive predictive value of the test}

\item{NPV}{numeric value representing the negative predictive value of the test}
}
\value{
A list containing the test results with attributes of class "test_results"
}
\description{
This function takes in four numeric values representing the sensitivity, specificity, positive predictive value (PPV), and negative predictive value (NPV) of a test, and returns a list containing the test results with attributes of class "test_results".
}
