% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Youden_Index.R
\name{youden}
\alias{youden}
\title{Calculate Youden's index}
\usage{
youden(sensitivity, specificity)
}
\arguments{
\item{sensitivity}{A numeric value representing the sensitivity of a test.}

\item{specificity}{A numeric value representing the specificity of a test.}
}
\value{
A Youden object with class "youden" and attributes sensitivity and specificity.
}
\description{
This function calculates Youden's index given sensitivity and specificity values.
}
\references{
Youden, W. J. (1950). Index for rating diagnostic tests. Cancer, 3(1), 32-35.
}
